Restructure the Compact runtime libraries
- `compact-types.ts` holds types emitted by the compiler (like `JubjubPoint` and
  `Secp256k1Point`) and the base built-in descriptors used by the compiler to
  build all emitted descriptors.  The `toBinaryRepr` helper stays there because
  it has to do with FAB.  Everything else is either deleted because it is unused
  (and duplicates the compiler generated descriptors) or is moved elsewhere if
  it's needed for some other part of the runtime.

- `built-ins.ts` holds functions emitted by the compiler in `print-typescript`
  and to implement the Midnight natives.  Utility-style functions have been
  moved to `utils.ts`.

This is a breaking change because some unnecessary and unused exported types and
descriptors are removed, and some exported helpers are no longer exported
because they're moved to where they're needed.
