+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Toolchain 0.33.111, language 0.25.103, runtime 0.18.103]
+
+### Changed
+
+- Clean up the Compact runtime to reflect the intended structure: types and
+  descriptors needed by the generated code are in `compact-types.ts`, but not
+  redundant and unnecessary implementations; functions used by emitted code are
+  in `built-ins.ts`, but not helpful utility functions; those are in `utils.ts`.
+
+### Internal notes
+
+- This is a breaking change because some unused and unnecessary exported types
+  and descriptors have been deleted.
+
 ## [Toolchain 0.33.111, language 0.25.103, runtime 0.18.102]
 
 ### Changed
