+++ runtime/src/built-ins.ts
@@ -21,11 +21,10 @@ import {
   CompactType,
   CompactTypeJubjubPoint,
   JubjubPoint,
-  JubjubSchnorrSignature,
   Secp256k1Point,
   toBinaryRepr,
 } from './compact-types.js';
-import { convertNumericToJubjubScalar } from './casts.js';
+import { secp256k1FromProjective, secp256k1ToProjective } from './utils.js';
 import { CompactError } from './error.js';
 
 /**
@@ -374,33 +373,6 @@ export function secp256k1PointX(pt: Secp256k1Point): bigint {
 export function secp256k1PointY(pt: Secp256k1Point): bigint {
   return pt.y;
 }
-
-/**
- * Lift the simple affine `Secp256k1Point` representation into a noble-curves
- * projective point. Identity maps to `Point.ZERO`; every other input is validated
- * to lie on the curve by `fromAffine`.
- */
-function secp256k1ToProjective(p: Secp256k1Point): ReturnType<typeof secp256k1.Point.fromAffine> {
-  if (p.identity) {
-    return secp256k1.Point.ZERO;
-  }
-  return secp256k1.Point.fromAffine({ x: p.x, y: p.y });
-}
-
-/**
- * Project a noble-curves point back down to the simple affine
- * `Secp256k1Point` representation.
- */
-function secp256k1FromProjective(p: ReturnType<typeof secp256k1.Point.fromAffine>): Secp256k1Point {
-  const k = p.toAffine();
-  if (/* k == secp256k1.Point.ZERO */ k.x == 0n && k.y == 0n) {
-    return { x: 0n, y: 0n, identity: true };
-  } else {
-    const { x, y } = k;
-    return { x: x, y: y, identity: false };
-  }
-}
-
 /**
  * The Compact builtin `ecAdd` function for secp256k1 points.
  *
@@ -433,34 +405,6 @@ export function secp256k1MulGenerator(b: bigint): Secp256k1Point {
 }
 
 /**
- * Recover the secp256k1 public key from an ECDSA signature and a message hash.
- *
- * ## Recovery ID
- * - bit 0 (`recoveryId & 1`) is the parity of `R.y`: 0 for even, 1 for odd.
- * - bit 1 (`recoveryId >= 2`) says whether the reduction wrapped, i.e. whether
- *   `R.x` is `r` (0, 1) or `r + n` (2, 3).
- *
- * - 0: `R = (r, y)` with `y` even — the common case.
- * - 1: `R = (r, y)` with `y` odd — the other common case.
- * - 2: `R = (r + n, y)` with `y` even.
- * - 3: `R = (r + n, y)` with `y` odd.
- */
-export function secp256k1EcdsaRecover(
-  msgHash: Uint8Array,
-  sig: { readonly r: bigint; readonly s: bigint },
-  recoveryId: number,
-): Secp256k1Point {
-  if (msgHash.length !== 32) {
-    throw new CompactError('expected a 32-byte message hash');
-  }
-  if (!Number.isInteger(recoveryId) || recoveryId < 0 || recoveryId > 3) {
-    throw new CompactError('expected a recovery id in the range [0, 3]');
-  }
-  const nobleSig = new secp256k1.Signature(sig.r, sig.s, recoveryId);
-  return secp256k1FromProjective(nobleSig.recoverPublicKey(msgHash));
-}
-
-/**
  * Concatenates multiple {@link AlignedValue}s
  * @internal
  */
@@ -471,95 +415,4 @@ export function alignedConcat(...values: ocrt.AlignedValue[]): ocrt.AlignedValue
     res.alignment = res.alignment.concat(value.alignment);
   }
   return res;
-}
-
-/**
- * Samples a random JubJub scalar.
- *
- * The returned value is in the range [0, JUBJUB_SCALAR_MODULUS).
- */
-export function jubjubSampleScalar(): bigint {
-  return ocrt.valueToBigInt(ocrt.jubjubSampleScalar());
-}
-
-/**
- * Alias for {@link jubjubSampleScalar}. Samples a random JubJub Schnorr signing key.
- */
-export const sampleJubjubSchnorrSk = jubjubSampleScalar;
-
-/**
- * Derives the Schnorr verifying key (public key) from a signing key.
- *
- * Equivalent to {@link ecMulGenerator}(signingKey).
- */
-export function jubjubSchnorrVerifyingKey(signingKey: bigint): JubjubPoint {
-  return ecMulGenerator(convertNumericToJubjubScalar(signingKey));
-}
-
-/**
- * Produces a Schnorr signature over the JubJub curve.
- *
- * - `rtType` / `msg`: the message as a typed Compact value
- * - `sk`: signing key as a JubJub scalar (e.g. as returned by {@link jubjubSampleScalar})
- *
- * The signature scheme:
- * - Nonce `r` sampled uniformly at random
- * - Announcement `R = r·G`
- * - Challenge `c = PoseidonHash(R.x, R.y, pk.x, pk.y, msg...)`
- * - Response `s = r + c·sk` (in the JubJub scalar field)
- */
-export function jubjubSchnorrSign<A>(rtType: CompactType<A>, msg: A, signingKey: bigint): JubjubSchnorrSignature {
-  const r = jubjubSampleScalar();
-  const announcement = ecMulGenerator(r);
-  const verifyingKey = ecMulGenerator(signingKey);
-
-  const challengeAlignment: ocrt.Alignment = [
-    ...CompactTypeJubjubPoint.alignment(),
-    ...CompactTypeJubjubPoint.alignment(),
-    ...rtType.alignment(),
-  ];
-  const challengeValue: ocrt.Value = [
-    ...CompactTypeJubjubPoint.toValue(announcement),
-    ...CompactTypeJubjubPoint.toValue(verifyingKey),
-    ...rtType.toValue(msg),
-  ];
-  const c = convertNumericToJubjubScalar(ocrt.valueToBigInt(ocrt.transientHash(challengeAlignment, challengeValue)));
-
-  const response = convertNumericToJubjubScalar(r + c * signingKey);
-  return { announcement, response };
-}
-
-/**
- * Verifies a Schnorr signature over the JubJub curve.
- *
- * - `rtType` / `msg`: the message as a typed Compact value
- * - `pk`: verifying key (a JubJubPoint / EmbeddedGroupAffine)
- * - `sig`: signature as returned by {@link jubjubSchnorrSign}
- *
- * Returns `true` if the signature is valid (i.e. `s·G == R + c·pk`).
- */
-export function jubjubSchnorrVerify<A>(
-  rtType: CompactType<A>,
-  msg: A,
-  verifyingKey: JubjubPoint,
-  sig: JubjubSchnorrSignature,
-): boolean {
-  const { announcement, response } = sig;
-
-  const challengeAlignment: ocrt.Alignment = [
-    ...CompactTypeJubjubPoint.alignment(),
-    ...CompactTypeJubjubPoint.alignment(),
-    ...rtType.alignment(),
-  ];
-  const challengeValue: ocrt.Value = [
-    ...CompactTypeJubjubPoint.toValue(announcement),
-    ...CompactTypeJubjubPoint.toValue(verifyingKey),
-    ...rtType.toValue(msg),
-  ];
-  const c = convertNumericToJubjubScalar(ocrt.valueToBigInt(ocrt.transientHash(challengeAlignment, challengeValue)));
-
-  const lhs = ecMulGenerator(response);
-  const rhs = ecAdd(announcement, ecMul(verifyingKey, c));
-
-  return lhs.x === rhs.x && lhs.y === rhs.y;
 }
